requirements.txt

streamlit==1.39.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy
requests
huggingface_hub==0.14.1
